+++ packages/fiber/__mocks__/react-native/Libraries/Pressability/Pressability.ts
@@ -1,4 +1,0 @@
-export default class {
-  getEventHandlers = () => ({})
-  reset() {}
-}