+++ packages/fiber/src/native.tsx
@@ -4,3 +4,6 @@ export * from './three-types'
 export * from './core'
 export * from './native/Canvas'
 export { createTouchEvents as events } from './native/events'
+
+import { polyfills } from './native/polyfills'
+polyfills()