+++ packages/fiber/tests/hooks.test.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react'
 import * as THREE from 'three'
 import * as Stdlib from 'three-stdlib'
+import { createCanvas } from '@react-three/test-renderer/src/createTestCanvas'
+
 import {
   createRoot,
   advance,
@@ -17,6 +19,12 @@ import {
 const root = createRoot(document.createElement('canvas'))
 
 describe('hooks', () => {
+  let canvas: HTMLCanvasElement = null!
+
+  beforeEach(() => {
+    canvas = createCanvas()
+  })
+
   it('can handle useThree hook', async () => {
     let result = {} as {
       camera: THREE.Camera
@@ -113,25 +121,21 @@ describe('hooks', () => {
     mesh2.name = 'Mesh 2'
     MockGroup.add(mesh1, mesh2)
 
-    jest.spyOn(Stdlib, 'GLTFLoader').mockImplementation(
-      () =>
-        ({
-          load: jest
-            .fn()
-            .mockImplementationOnce((_url, onLoad) => {
-              onLoad(MockMesh)
-            })
-            .mockImplementationOnce((_url, onLoad) => {
-              onLoad({ scene: MockGroup })
-            }),
-          setPath: () => {},
-        } as unknown as Stdlib.GLTFLoader),
-    )
+    class TestLoader extends THREE.Loader {
+      load = jest
+        .fn()
+        .mockImplementationOnce((_url, onLoad) => {
+          onLoad(MockMesh)
+        })
+        .mockImplementationOnce((_url, onLoad) => {
+          onLoad(MockGroup)
+        })
+    }
+
+    const extensions = jest.fn()
 
     const Component = () => {
-      const [mockMesh, mockScene] = useLoader(Stdlib.GLTFLoader, ['/suzanne.glb', '/myModels.glb'], (loader) => {
-        loader.setPath('/public/models')
-      })
+      const [mockMesh, mockScene] = useLoader(TestLoader, ['/suzanne.glb', '/myModels.glb'], extensions)
 
       return (
         <>
@@ -145,16 +149,20 @@ describe('hooks', () => {
     const { scene } = store.getState()
 
     expect(scene.children[0]).toBe(MockMesh)
+    expect(scene.children[1]).toBe(MockGroup)
+    expect(extensions).toBeCalledTimes(1)
   })
 
   it('can handle useLoader with a loader extension', async () => {
     class Loader extends THREE.Loader {
-      load = (_url: string) => null
+      load(_url: string, onLoad: (result: null) => void): void {
+        onLoad(null)
+      }
     }
 
     let proto!: Loader
 
-    function Test() {
+    function Test(): null {
       return useLoader(Loader, '', (loader) => (proto = loader))
     }
     await act(async () => root.render(<Test />))