+++ packages/fiber/tests/index.test.tsx
@@ -2,8 +2,17 @@ import * as React from 'react'
 import * as THREE from 'three'
 import * as ts from 'typescript'
 import * as path from 'path'
-import { ReconcilerRoot, createRoot, act, useFrame, useThree, createPortal, RootState, RootStore } from '../src/index'
-import { privateKeys } from '../src/core/store'
+import {
+  ReconcilerRoot,
+  createRoot,
+  act,
+  useFrame,
+  useThree,
+  createPortal,
+  RootState,
+  RootStore,
+  extend,
+} from '../src/index'
 
 let root: ReconcilerRoot<HTMLCanvasElement> = null!
 
@@ -118,36 +127,103 @@ describe('createRoot', () => {
     expect(gl.physicallyCorrectLights).toBe(true)
   })
 
+  it('should update scene via scene prop', async () => {
+    let scene: THREE.Scene = null!
+
+    await act(async () => {
+      scene = root
+        .configure({ scene: { name: 'test' } })
+        .render(<group />)
+        .getState().scene
+    })
+
+    expect(scene.name).toBe('test')
+  })
+
+  it('should set a custom scene via scene prop', async () => {
+    let scene: THREE.Scene = null!
+
+    const prop = new THREE.Scene()
+
+    await act(async () => {
+      scene = root
+        .configure({ scene: prop })
+        .render(<group />)
+        .getState().scene
+    })
+
+    expect(prop).toBe(scene)
+  })
+
   it('should set a renderer via gl callback', async () => {
     class Renderer extends THREE.WebGLRenderer {}
 
-    const store = await act(async () => root.configure({ gl: (canvas) => new Renderer({ canvas }) }).render(<group />))
-    const { gl } = store.getState()
+    let gl: Renderer = null!
+    await act(async () => {
+      gl = root
+        .configure({ gl: (canvas) => new Renderer({ canvas }) })
+        .render(<group />)
+        .getState().gl
+    })
 
     expect(gl instanceof Renderer).toBe(true)
   })
 
   it('should respect color management preferences via gl', async () => {
-    const store = await act(async () =>
-      root
-        .configure({ gl: { outputEncoding: THREE.LinearEncoding, toneMapping: THREE.NoToneMapping } })
-        .render(<group />),
-    )
-    const { gl } = store.getState()
+    const texture = new THREE.Texture() as THREE.Texture & { colorSpace?: string }
+    let key = 0
+    function Test() {
+      return <meshBasicMaterial key={key++} map={texture} />
+    }
 
-    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
-    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
+    const LinearEncoding = 3000
+    const sRGBEncoding = 3001
 
-    await act(async () => root.configure({ flat: true, linear: true }).render(<group />))
-    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
+    let gl: THREE.WebGLRenderer & { outputColorSpace?: string } = null!
+    await act(async () => (gl = root.render(<Test />).getState().gl))
+    expect(gl.outputEncoding).toBe(sRGBEncoding)
+    expect(gl.toneMapping).toBe(THREE.ACESFilmicToneMapping)
+    expect(texture.encoding).toBe(sRGBEncoding)
+
+    await act(async () => root.configure({ linear: true, flat: true }).render(<Test />))
+    expect(gl.outputEncoding).toBe(LinearEncoding)
     expect(gl.toneMapping).toBe(THREE.NoToneMapping)
+    expect(texture.encoding).toBe(LinearEncoding)
+
+    // Sets outputColorSpace since r152
+    const SRGBColorSpace = 'srgb'
+    const LinearSRGBColorSpace = 'srgb-linear'
+
+    gl.outputColorSpace ??= ''
+    texture.colorSpace ??= ''
+
+    await act(async () => root.configure({ linear: true }).render(<Test />))
+    expect(gl.outputColorSpace).toBe(LinearSRGBColorSpace)
+    expect(texture.colorSpace).toBe(LinearSRGBColorSpace)
+
+    await act(async () => root.configure({ linear: false }).render(<Test />))
+    expect(gl.outputColorSpace).toBe(SRGBColorSpace)
+    expect(texture.colorSpace).toBe(SRGBColorSpace)
   })
 
   it('should respect legacy prop', async () => {
-    await act(async () => root.configure({ legacy: true }).render(<group />))
+    // <= r138 internal fallback
+    const material = React.createRef<THREE.MeshBasicMaterial>()
+    extend({ ColorManagement: null! })
+    await act(async () => root.render(<meshBasicMaterial ref={material} color="#111111" />))
+    expect((THREE as any).ColorManagement.legacyMode).toBe(false)
+    expect(material.current!.color.toArray()).toStrictEqual(new THREE.Color('#111111').convertSRGBToLinear().toArray())
+    extend({ ColorManagement: (THREE as any).ColorManagement })
+
+    // r139 legacyMode
+    await act(async () => {
+      root.configure({ legacy: true }).render(<group />)
+    })
     expect((THREE as any).ColorManagement.legacyMode).toBe(true)
 
-    await act(async () => root.configure({ legacy: false }).render(<group />))
+    await act(async () => {
+      root.configure({ legacy: false }).render(<group />)
+    })
     expect((THREE as any).ColorManagement.legacyMode).toBe(false)
 
     // r150 !enabled
@@ -201,11 +277,6 @@ describe('createPortal', () => {
     // Creates an isolated state enclave
     expect(state.scene).not.toBe(scene)
     expect(portalState.scene).toBe(scene)
-
-    // Preserves internal keys
-    const overwrittenKeys = ['get', 'set', 'events', 'size', 'viewport']
-    const respectedKeys = privateKeys.filter((key) => overwrittenKeys.includes(key) || state[key] === portalState[key])
-    expect(respectedKeys).toStrictEqual(privateKeys)
   })
 
   it('should handle unmounted containers', async () => {