+++ packages/fiber/tests/renderer.test.tsx
@@ -3,15 +3,22 @@ import * as THREE from 'three'
 import { ReconcilerRoot, createRoot, act, extend, ThreeElement } from '../src/index'
 import { suspend } from 'suspend-react'
 
-class CustomElement extends THREE.Object3D {}
-
-declare module '@react-three/fiber' {
-  interface ThreeElements {
-    customElement: ThreeElement<typeof CustomElement>
+class Mock extends THREE.Group {
+  static instances: string[]
+  constructor(name: string = '') {
+    super()
+    this.name = name
+    Mock.instances.push(name)
   }
 }
 
-extend({ CustomElement })
+declare module '@react-three/fiber' {
+  interface ThreeElements {
+    mock: ThreeElement<typeof Mock>
+  }
+}
+
+extend({ Mock })
 
 type ComponentMesh = THREE.Mesh<THREE.BoxBufferGeometry, THREE.MeshBasicMaterial>
 
@@ -34,7 +41,10 @@ const expectToThrow = async (callback: () => any) => {
 describe('renderer', () => {
   let root: ReconcilerRoot<HTMLCanvasElement> = null!
 
-  beforeEach(() => (root = createRoot(document.createElement('canvas'))))
+  beforeEach(() => {
+    root = createRoot(document.createElement('canvas'))
+    Mock.instances = []
+  })
   afterEach(async () => act(async () => root.unmount()))
 
   it('should render empty JSX', async () => {
@@ -45,35 +55,32 @@ describe('renderer', () => {
   })
 
   it('should render native elements', async () => {
-    const store = await act(async () => root.render(<group />))
+    const store = await act(async () => root.render(<group name="native" />))
     const { scene } = store.getState()
 
     expect(scene.children.length).toBe(1)
     expect(scene.children[0]).toBeInstanceOf(THREE.Group)
+    expect(scene.children[0].name).toBe('native')
   })
 
   it('should render extended elements', async () => {
-    const store = await act(async () => root.render(<customElement />))
+    const store = await act(async () => root.render(<mock name="mock" />))
     const { scene } = store.getState()
 
     expect(scene.children.length).toBe(1)
-    expect(scene.children[0]).toBeInstanceOf(CustomElement)
-
-    const Component = extend(THREE.Mesh)
-    await act(async () => root.render(<Component />))
-
-    expect(scene.children.length).toBe(1)
-    expect(scene.children[0]).toBeInstanceOf(THREE.Mesh)
+    expect(scene.children[0]).toBeInstanceOf(Mock)
+    expect(scene.children[0].name).toBe('mock')
   })
 
   it('should render primitives', async () => {
     const object = new THREE.Object3D()
 
-    const store = await act(async () => root.render(<primitive object={object} />))
+    const store = await act(async () => root.render(<primitive name="primitive" object={object} />))
     const { scene } = store.getState()
 
     expect(scene.children.length).toBe(1)
     expect(scene.children[0]).toBe(object)
+    expect(object.name).toBe('primitive')
   })
 
   it('should go through lifecycle', async () => {
@@ -455,9 +462,9 @@ describe('renderer', () => {
       suspend(async (reconstruct) => reconstruct, [reconstruct])
 
       return (
-        <group key={reconstruct ? 0 : 1} name="parent">
-          <group
-            name="child"
+        <mock key={reconstruct ? 0 : 1} args={['parent']}>
+          <mock
+            args={['child']}
             ref={(self) => void (lastMounted = self?.uuid)}
             attach={(_, self) => {
               calls.push('attach')
@@ -465,7 +472,7 @@ describe('renderer', () => {
               return () => calls.push('detach')
             }}
           />
-        </group>
+        </mock>
       )
     }
 
@@ -473,9 +480,9 @@ describe('renderer', () => {
       React.useLayoutEffect(() => void calls.push('useLayoutEffect'), [])
 
       return (
-        <group name="suspense">
+        <mock args={['suspense']}>
           <SuspenseComponent {...props} />
-        </group>
+        </mock>
       )
     }
 
@@ -484,10 +491,12 @@ describe('renderer', () => {
     // Should complete tree before layout-effects fire
     expect(calls).toStrictEqual(['attach', 'useLayoutEffect'])
     expect(lastAttached).toBe(lastMounted)
+    expect(Mock.instances).toStrictEqual(['suspense', 'parent', 'child'])
 
     await act(async () => root.render(<Test reconstruct />))
 
     expect(calls).toStrictEqual(['attach', 'useLayoutEffect', 'detach', 'attach'])
     expect(lastAttached).toBe(lastMounted)
+    expect(Mock.instances).toStrictEqual(['suspense', 'parent', 'child', 'parent', 'child'])
   })
 })